+++ include/kite/update.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace kite {
+
+int run_update(bool check_only, bool force);
+
+} // namespace kite